Update README and refactor utility functions
- Enhance README with project description and demo
- Rename command_utils.ts to prompt_utils.ts
- Move INIT_INSTRUCTIONS constant to prompt_utils.ts
- Update import statements across multiple files
